serveur/requirements.txt
# pip install -r requirements.txt
Flask
flask-cors
llama-cpp-python